feat: refine distance calculations and add subtraction methods for fixed-step domains

+++ src/Domain/Intervals.NET.Domain.Default/Numeric/IntegerFixedStepDomain.cs
@@ -22,7 +22,7 @@ public readonly struct IntegerFixedStepDomain : IFixedStepDomain<int>
     /// <inheritdoc/>
     [Pure]
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public long Distance(int start, int end) => end - start;
+    public long Distance(int start, int end) => (long)end - (long)start;
 
     /// <inheritdoc/>
     [Pure]
